feat: sharedモジュールからdataモジュールへクラスを移動させた
