+++ androidApp/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/MainApplication.kt
@@ -3,7 +3,7 @@ package com.ikmr.banbara23.yaeyama_liner_checker
 import android.app.Application
 import com.google.firebase.crashlytics.BuildConfig
 import com.google.firebase.crashlytics.FirebaseCrashlytics
-import com.ikemura.shared.di.initKoin
+import data.di.initKoin
 import timber.log.Timber
 
 class MainApplication : Application() {
