+++ androidApp/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/ui/portstatusdetail/PortStatusDetailViewModel.kt
@@ -1,8 +1,8 @@
 package com.ikmr.banbara23.yaeyama_liner_checker.ui.portstatusdetail
 
 import androidx.lifecycle.ViewModel
-import com.ikemura.shared.repository.StatusDetailRepository
-import com.ikemura.shared.usecase.GetStatusDetail
+import data.repository.StatusDetailRepository
+import data.usecase.GetStatusDetail
 
 /**
  * 運行詳細 ViewModel
